% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECCO.R
\name{ecco}
\alias{ecco}
\title{Controlling for Confounding Effects in eQTL Mapping Studies through Joint Differential Expression and Mendelian Randomization Analyses}
\usage{
ecco(pheno, gene, gene_name, iv_snp, peer)
}
\arguments{
\item{pheno}{the phenotype data}

\item{gene}{a matrix containing the whole gene expression data after removing the PEER factors}

\item{gene_name}{a vector containg the names of all genes}

\item{iv_snp}{the instrumental variable, produced from ecco0}

\item{peer}{the number of peer factors to be examined}
}
\value{
\item{Gene}{the name of the gene}

\item{PEER}{the number of PEER factors}

\item{p-value}{the p-value of alpha}

\item{beta_hat}{the estimation of beta from MR model}

\item{beta_tilde}{the estimation of beta from DE model}
}
\description{
Determining the optimal number of PEER factors for eQTL mapping analysis through DE and MR analysis
}
\details{
Instead of performing repetitive eQTL mapping, ECCO jointly applies differential expression analysis and Mendelian randomization (MR) analysis, leading to substantial computational savings.
}
\examples{
data(exampledata)
attach(exampledata)
num_peer=1
 summary<-ecco(pheno,peer[[num_peer]],gene_name,iv_snp,num_peer)
closeAllConnections()
detach(exampledata)
}
