% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECCO.R
\name{ecco0}
\alias{ecco0}
\title{Controlling for Confounding Effects in eQTL Mapping Studies through Joint Differential Expression and Mendelian Randomization Analyses}
\usage{
ecco0(gene, genename, gene_name, geno, ind)
}
\arguments{
\item{gene}{the gene expression data.}

\item{genename}{the name of the gene}

\item{gene_name}{a vector containg the names of all genes}

\item{geno}{a matrix containg all the cis-SNPs of the analyzed gene}

\item{ind}{the index of the gene}
}
\value{
\item{iv_snp}{the snp that has the strongest association with the gene}
}
\description{
Determining the optimal number of PEER factors for eQTL mapping analysis through DE and MR analysis
}
\details{
Instead of performing repetitive eQTL mapping, ECCO jointly applies differential expression analysis and Mendelian randomization (MR) analysis, leading to substantial computational savings.
}
\examples{
data(exampledata)
attach(exampledata)
ind=1
genename=gene_name[ind]
gene=M_matrix[,ind]
geno=snp_raw[[ind]]
ivsnp=ecco0(gene,genename,gene_name,geno,ind)
closeAllConnections()
detach(exampledata)
}
\author{
Yue Fan, Shiquan Sun, Xiang Zhou
}
