% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{Example dataset}
\format{Contains the following objects:
\describe{
  \item{gene_name}{a vector containg the names of all genes.}
  \item{M_matrix}{a matrix containing the whole gene expression data.}
  \item{peer}{a data list containing the whole gene expression data after removing the PEER factors.}
  \item{snp_raw}{a data list containing the whole genotype data.}
  \item{Y}{the phenotype data.}
  \item{peerlist}{the number of peer factors to be examined.}
}}
\usage{
data(exampledata)
}
\description{
A simulated example dataset of ECCO.
The variables are as follows:
}
\keyword{datasets}
